Excludes ghpages files from headers changes
